Don't put a dot in front of *.c.dep in order to avoid a make issue
It turns out that GNU make's implementation of multiple targets has a
serious issue.  If you do something like:

foo.o .foo.c.dep: foo.c
	...recipe...

GNU will internally expand this as two distinct rules for the two
targets, which can result in the recipe being invoked twice and incorrect
parallel builds. But there is an exception for pattern rules.  If you do:

%.o %.c.dep: %.c
	...recipe...

GNU make will understand that this is a single rule with two targets.

(Hat tip: http://blog.melski.net/2013/01/01/pragma-multi-and-rules-with-multiple-outputs-in-gnu-make/)

So we need the main .c rule to be a pattern rule.  But our rule for compiling
.c files was not a pure pattern rule.  It was:

%.o $(call dotify,%.c.dep): %.c

It seems that "$(call dotify,%.c.dep)" falls foul of the multiple target
issue, despite containing "%".  So we need to drop the use of the dotify
function, which is there to add a dot in front of the filename in the
presence of subdirectories.  So we can no longer do that.