Replace the eu_parse buffer with a scratch area in the stack region
Because we already manage the use of the stack region and resize it
when necessary, it's fairly easy to reserve a "scratch" area at the
start for use by the outermost stack frame.  This can replace the
use of the buffer, reducing allocations and making struct eu_parse
slightly smaller.